{{cookiecutter.project_name}}/requirements.txt
adal==0.4.7
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
azure==2.0.0
azure-batch==3.0.0
azure-common==1.1.8
azure-datalake-store==0.0.17
azure-graphrbac==0.30.0
azure-keyvault==0.3.7
azure-mgmt==1.0.0
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==4.0.0
azure-mgmt-cdn==0.30.3
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==1.0.0
azure-mgmt-containerregistry==0.2.1
azure-mgmt-datalake-analytics==0.1.6
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.6
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.31.0
azure-mgmt-logic==2.1.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==4.1.0
azure-mgmt-resource==1.1.0
azure-mgmt-scheduler==1.1.3
azure-mgmt-sql==0.5.3
azure-mgmt-storage==1.0.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==5.6.130
azure-servicemanagement-legacy==0.20.6
azure-storage==0.20.0
certifi==2017.7.27.1
cffi==1.11.2
channels==1.1.8
channels-api==0.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.1
daphne==1.4.2
Django==1.11.6
django-appconf==1.0.2
django-braces==1.11.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-imagekit==4.0.1
django-oauth-toolkit==1.0.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.0.8
django-storages==1.6.5
djangorestframework==3.7.7
enum34==1.1.6
futures==3.1.1
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
isodate==0.6.0
keyring==10.4.0
msgpack-python==0.5.4
msrest==0.4.17
msrestazure==0.4.15
oauthlib==2.0.6
olefile==0.44
pathlib2==2.3.0
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
SecretStorage==2.3.1
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
whitenoise==3.3.1
zope.interface==4.4.3
